create distributed.py
