+++ homura/__init__.py
@@ -1,8 +1,7 @@
 from .register import Registry
-from .utils.containers import TensorDataClass, TensorTuple
-from .utils.environment import enable_accimage, get_args, get_environ, get_git_hash, get_global_rank, get_local_rank, \
-    get_num_nodes, get_world_size, if_is_master, init_distributed, is_accimage_available, is_distributed, \
-    is_distributed_available, is_faiss_available, is_horovod_available, is_master
-from .utils.reproducibility import set_deterministic, set_seed
+from .utils import TensorDataClass, TensorTuple, enable_accimage, get_args, get_environ, get_git_hash, get_global_rank, \
+    get_local_rank, get_num_nodes, get_world_size, if_is_master, init_distributed, is_accimage_available, \
+    is_distributed, is_distributed_available, is_faiss_available, is_horovod_available, is_master, set_deterministic, \
+    set_seed
 
 Registry.import_modules('homura.vision')
