+++ homura/utils/__init__.py
@@ -1,6 +1,7 @@
 from .benchmarks import timeit
 from .containers import TensorDataClass, TensorTuple
-from .environment import (enable_accimage, get_args, get_environ, get_git_hash, get_global_rank, get_local_rank,
-                          get_num_nodes, get_world_size, init_distributed, is_accimage_available, is_distributed,
-                          is_distributed_available, is_faiss_available, is_horovod_available, is_master)
+from .distributed import (get_global_rank, get_local_rank, get_num_nodes, get_world_size, if_is_master,
+                          init_distributed, is_distributed, is_distributed_available, is_horovod_available, is_master)
+from .environment import (enable_accimage, get_args, get_environ, get_git_hash, is_accimage_available,
+                          is_faiss_available)
 from .reproducibility import set_deterministic, set_seed
